pyproject.toml
[tool.poetry]
name = "roster-cli"
version = "0.1.0"
description = "Local file-driven constraint rostering engine for cricket, church, and on-call schedules"
authors = ["Roster Team"]
readme = "README.md"
packages = [{include = "roster_cli"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
typer = {extras = ["all"], version = "^0.9.0"}
rich = "^13.7.0"
pydantic = "^2.5.0"
pyyaml = "^6.0.1"
ics = "^0.7.2"
python-dateutil = "^2.8.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
ruff = "^0.1.7"
black = "^23.12.0"
mypy = "^1.7.1"
types-pyyaml = "^6.0.12"
types-python-dateutil = "^2.8.19"
pyinstaller = "^6.3.0"

[tool.poetry.scripts]
roster = "roster_cli.main:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.black]
line-length = 100
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = "ics.*"
ignore_missing_imports = true
